==> CODE/00a_writing_frames_for_annotation.ipynb <==
from subprocess import call

call(cd, C:\Users\cloftus\Desktop\)


import fnv
import fnv.reduce
import fnv.file  
import numpy as np
import seaborn as sns
import datetime as dt
import time
import pandas as pd
from tkinter import filedialog
from tkinter import *
import os
import matplotlib.pyplot as plt
import glob
import h5py
import cv2

h5py.run_tests()

os.chdir('C:/Users/cloftus/Desktop/')

options = np.array([000000, 200000, 400000, 600000, 800000, 1000000])

def round_to_fifth(time_):
    
    new = options[ abs(time_.microsecond - options).argmin() ]
    
    if new != 1000000:
        new_time = time_.replace(microsecond = new)
    else:
        temp_time = time_ + dt.timedelta(seconds = 1)
        new_time = temp_time.replace(microsecond = 000000)
    
    return(new_time)

def time_correct( _timestamp_ ):
    if drive == 'D:/thermal_data/':
        _timestamp_ = _timestamp_ - time_offset_D
    if drive == 'F:/':
        _timestamp_ = _timestamp_ - time_offset_F
        
    return(_timestamp_)

# root = Tk()
# root.filename =  filedialog.askopenfilename(initialdir = "/",title = "Select file",filetypes = (("jpeg files","*.jpg"),("all files","*.*")))
# print (root.filename)

datelist = [ d.date() for d in pd.date_range(start = pd.to_datetime('2019-08-06'), end = pd.to_datetime('2019-08-07')).to_pydatetime()]
print(datelist)

## writing only subsets of videos for annotating reading from hard drives.
## pulls 200 frames that are evenly distributed in time throughout the night

zoom = 5

frames_per_night = 200

# positive offset implies that the given video is delayed compared to the T1020 video
# negative offset implies that the given video is ahead of the T1020 video
# note that I never actually used the offset because I only used the T1020 videos
time_offset_D = 0

time_offset_F = 0

for date in datelist:
    
    curr_date = f'{date.year:04}' + f'{date.month:02}' + f'{date.day:02}'
    
    start_times = []
    end_times = []
    
    assign_list = []
    
    drives = ['E:/', 'D:/thermal_data/', 'F:/']
    
    for drive in drives:
        
        if os.path.isdir(drive + curr_date) == True:
            
            os.chdir(drive + curr_date)
            file_name = glob.glob('*.seq')
            print(file_name, flush = True)
            
            if len(file_name) > 0:
                
                file_assign = 'im_' + drive[0]
                
                start_ = time.time()
                
                globals()[file_assign] = fnv.file.ImagerFile(file_name[0])
                
                print( 'loading: ', time.time() - start_, flush = True )
                
                assign_list.append( file_assign )
                
                im = globals()[file_assign]
                
                im.get_frame(0)
                day1 = int(im.frame_info[0]['value'][:3])
                start = f'{date.year:04}-' + f'{date.month:02}-' + f'{date.day:02} ' + im.frame_info[0]['value'][4:]
                start_dt = dt.datetime.strptime(start, '%Y-%m-%d %H:%M:%S.%f')
                start_dt = time_correct(start_dt)
                start_seq = round_to_fifth(start_dt)
                start_times.append(start_seq)
                
                print( 'checkpoint 1', flush = True )

                im.get_frame(im.num_frames - 1)
                day2 = int(im.frame_info[0]['value'][:3])
                day_diff = day2 - day1
                end = f'{date.year:04}-' + f'{date.month:02}-' + f'{(date.day+day_diff):02} ' + im.frame_info[0]['value'][4:]
                end_dt = dt.datetime.strptime(end, '%Y-%m-%d %H:%M:%S.%f')
                end_dt = time_correct(end_dt)
                end_seq = round_to_fifth(end_dt)
                end_times.append(end_seq)

                print( 'checkpoint 2', flush = True )
                
                if len(file_name) > 1:
                    wehaveaproblem
    
    if len(start_times) == 0:
        continue
    
    start_all = np.array(start_times).min()
    end_all = np.array(end_times).max()
    
    step = (end_all - start_all) / frames_per_night
    
    time_list = []
    
    print( 'checkpoint 3', flush = True )
    
    x = 0
    
    last_time = start_all
    
    while last_time <= end_all:
        
        time_list.append(last_time)
        
        last_time += step 
        
    time_list = list( map( round_to_fifth, time_list ) )
        
    for assign in assign_list:
        
        start_ = time.time()

        im = globals()[assign]
        
        print( 'reloading: ', time.time() - start_, flush = True )

        chunk = 0

        im.get_frame(0)
        day1 = int(im.frame_info[0]['value'][:3])

        blank_mat = np.zeros((im.height, im.width))
        
        if im.width > 1000:  
            pre = 1
            
        else:
            if os.path.isdir('C:/Users/cloftus/Desktop/Results/' + '2_' + f'{date.year:04}' + f'{date.month:02}' + f'{date.day:02}') == False:
                pre = 2
            else:
                pre = 3

        os.makedirs( 'C:/Users/cloftus/Desktop/Results/' + str(pre) + '_' + f'{date.year:04}' + f'{date.month:02}' + f'{date.day:02}')

        os.chdir( 'C:/Users/cloftus/Desktop/Results/' + str(pre) + '_' + f'{date.year:04}' + f'{date.month:02}' + f'{date.day:02}')

        timestamps = []

        for i in range(im.num_frames):

            im.get_frame(i)

            day2 = int(im.frame_info[0]['value'][:3])

            day_diff = day2 - day1

            timestamp = f'{date.year:04}-' + f'{date.month:02}-' + f'{(date.day+day_diff):02} ' + im.frame_info[0]['value'][4:]

            timestamp = dt.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
            
            timestamp = time_correct( timestamp )

            timestamps.append([timestamp])


        rounded_timestamps = list(map( lambda x: round_to_fifth(x[0]), timestamps ))

        print( 'checkpoint 4', flush = True )

        for time_mark in time_list:

            ind = np.where( np.array(rounded_timestamps) == time_mark )

            if len(ind[0]) == 0:

                fig, ax = plt.subplots(figsize = (blank_mat.shape[1] / (100/zoom), blank_mat.shape[0] / (100/zoom)))  

                ax = sns.heatmap(blank_mat, xticklabels = False, yticklabels = False, cbar = False)
                
                if blank_mat.shape[1] > 1000:
                    
                    plt.xlabel( ind + time_mark.strftime(" %Y-%m-%d %H:%M:%S.%f") , fontsize = 55*zoom/5  )
                    
                else:
                    
                    plt.xlabel( ind + time_mark.strftime(" %Y-%m-%d %H:%M:%S.%f") , fontsize = 25*zoom/5  )

                plt.savefig( time_mark.strftime("%Y%m%d_%H%M%S%f") + '.tiff', bbox_inches = 'tight' )

                plt.close()

                print( 'not here', flush = True)

            else:

                im.get_frame( ind[0][0] )

                final = np.array(im.final, copy=False).reshape((im.height, im.width))

                final1 = final - final.mean()

                fig, ax = plt.subplots(figsize = (final.shape[1] / (100/zoom), final.shape[0] / (100/zoom)))  

                ax = sns.heatmap(final1, vmin = -np.std(final1), vmax = final1.max(), xticklabels = False, yticklabels = False, cbar = False)

                if blank_mat.shape[1] > 1000:
                    
                    plt.xlabel( ind + time_mark.strftime(" %Y-%m-%d %H:%M:%S.%f") , fontsize = 55*zoom / 5 )
                    
                else:
                    
                    plt.xlabel( ind + time_mark.strftime(" %Y-%m-%d %H:%M:%S.%f") , fontsize = 25*zoom / 5 )

                plt.savefig( time_mark.strftime("%Y%m%d_%H%M%S%f") + '.tiff' , bbox_inches = 'tight')

                plt.close()
                
                
        globals()[assign] = None

### helpful bit of code that renames files so they have a counter in front of the timestamp

root_path = 'C:/Users/cloftus/Desktop/Results/'

os.chdir(root_path)

folders = os.listdir()

for folder in folders:
        
        os.chdir(root_path + folder)
    
        files = sorted( os.listdir() )
    
        counter = 0
        
        for file in files:
            
            if 'tiff' in file:
                #print( file.split( '_', 1) )
                #print( file.split( '_', 1)[1] )
                #os.rename( file,   file.split( '_', 1)[1] )
                #os.rename( file,   file[4: ] )
                #os.rename( file,  f'{counter:05}' + '_' + file )
                
                counter += 1

## makes a txt file for each folder in Results/ of all of the tiff images in that folder 
## this txt file will be used by ffmpeg in the command line to string the pictures together into a video

root_path = 'C:/Users/cloftus/Desktop/Results/'

os.chdir(root_path)

folders = os.listdir()

for folder in folders:
        
    os.chdir(root_path + folder)
    
    files = sorted( os.listdir() )
    
    text_file = open(folder + '.txt', 'w')
        
    for file in files:

        if 'tiff' in file:

            n = text_file.write('file ' + "'" + file + "'" + '\n')

    text_file.close()

## command line script for turning the series of pictures produced in the script below to movies for upload
## this should be run in the command line (it won't run here)
cd C:\Users\cloftus\Desktop\Results\1_20190804
ffmpeg -f concat -r 30 -i 1_20190804.txt -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 1_20190804.mp4
