==> CODE/00c_writing_full_vids_hard_drives.ipynb <==
import fnv
import fnv.reduce
import fnv.file  
import numpy as np
import seaborn as sns
import datetime as dt
import time
import pandas as pd
from tkinter import filedialog
from tkinter import *
import os
import matplotlib.pyplot as plt
import glob
import gc
import shlex
import pipes
from subprocess import check_call

gc.collect()


options = np.array([000000, 200000, 400000, 600000, 800000, 1000000])

def round_to_fifth(time_):
    
    new = options[ abs(time_.microsecond - options).argmin() ]
    
    if new != 1000000:
        new_time = time_.replace(microsecond = new)
    else:
        temp_time = time_ + dt.timedelta(seconds = 1)
        new_time = temp_time.replace(microsecond = 000000)
    
    return(new_time)

def time_correct( _timestamp_ ):
    if drive == 'D:/thermal_data/':
        _timestamp_ = _timestamp_ - time_offset_D
    if drive == 'F:/':
        _timestamp_ = _timestamp_ - time_offset_F
        
    return(_timestamp_)

# root = Tk()
# root.filename =  filedialog.askopenfilename(initialdir = "/",title = "Select file",filetypes = (("jpeg files","*.jpg"),("all files","*.*")))
# print (root.filename)

datelist = [ d.date() for d in pd.date_range(start = pd.to_datetime('2019-08-11'), end = pd.to_datetime('2019-08-11')).to_pydatetime()]
print(datelist)

_date_list_ = [ '2019-08-03' ]

## writing all frames from a given thermal recording, separated into minute bins. Timestamps unrounded

zoom = 5

start_frame = 143369

vid_length = 1*60*60

drive = 'D:/'

start_timestamp = _date_list_[ 0 ]

write_to = 'E:/thermal_vids/' + start_timestamp

#write_to = 'L:/analysis/annotation_vids/' + start_timestamp

start_dt = dt.datetime.strptime(start_timestamp, '%Y-%m-%d')

curr_date = f'{start_dt.year:04}' + f'{start_dt.month:02}' + f'{start_dt.day:02}'

os.chdir( drive + curr_date )

file_name = glob.glob('*.seq')

im = fnv.file.ImagerFile(file_name[0])

im.get_frame(0)

print( 'got_frame')

day1 = int(im.frame_info[0]['value'][:3])

start = f'{start_dt.year:04}-' + f'{start_dt.month:02}-' + f'{start_dt.day:02} ' + im.frame_info[0]['value'][4:]

start_dt = dt.datetime.strptime(start, '%Y-%m-%d %H:%M:%S.%f')

og_start_dt = start_dt


if not os.path.isdir( write_to + '/' ):
    
    os.makedirs ( write_to + '/')

os.chdir( write_to + '/')


final = np.array(im.final, copy=False).reshape((im.height, im.width))
    
width = final.shape[1] / (100 / zoom)
    
height = final.shape[0] / (100 / zoom)

fig = plt.figure( figsize = ( width, height ) )  

if start_frame > 0:
    
    counter = 0
    
else:
    
    counter = 1
    
for i in range(start_frame, im.num_frames):
    
    if im.get_frame( i ) == False:
        blaasdfadff
        
    #start__ = time.time()

    im.get_frame(i)

    #print( 'get_frame: ' , time.time() - start__ )

    #start__ = time.time()
    
    day2 = int(im.frame_info[0]['value'][:3])

    day_diff = day2 - day1

    timestamp = f'{og_start_dt.year:04}-' + f'{og_start_dt.month:02}-' + f'{(og_start_dt.day+day_diff):02} ' + im.frame_info[0]['value'][4:]

    timestamp = dt.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
    
    #print( 'timestamp_info: ' , time.time() - start__ )

    #start__t = time.time()
    
    if timestamp > ( start_dt + dt.timedelta( seconds = vid_length ) ):
        
        if not counter == 0:
           
            str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

            # make a text_file of all the tiffs in the folder
            text_file = open(str_name + '.txt', 'w')

            files = sorted( os.listdir() )

            for file in files:

                if 'tiff' in file:

                    n = text_file.write('file ' + "'" + file + "'" + '\n')

            text_file.close()

            ## command line script for turning the series of pictures produced in the script below to movies for upload

            txt_name = str_name + '.txt'
            vid_name = str_name + '.mp4'

            command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
            check_call(shlex.split(command))

        
        to_delete = glob.glob('*.tiff')
        
        for item in to_delete:
            os.remove( item )
        
        start_dt = timestamp
        
        counter = 1
        #os.makedirs ( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
        
        #os.chdir( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
     
    
    #print( 'if_statement: ' , time.time() - start__t )

    #start__ = time.time()
        
    final = np.array(im.final, copy=False).reshape( ( im.height, im.width ) )

    final1 = final - final.mean()
    
    #print( 'matrix: ' , time.time() - start__ )

    #start__ = time.time()

    sns.heatmap(final1, vmin = -np.std(final1), vmax = final1.max(), xticklabels = False, yticklabels = False, cbar = False)

    if final1.shape[1] > 1000:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 55*zoom / 5 )

    else:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 25*zoom / 5 )

    
    #print( 'plotting: ' , time.time() - start__ )

    #start__ = time.time()

    plt.savefig( f'{i:06}_' + timestamp.strftime("%Y%m%d_%H%M%S%f") + '.tiff' , bbox_inches = 'tight')
    
    #print( 'save_plot: ' , time.time() - start__ )

    #start__ = time.time()
    
    fig.clear()
    
    #plt.clf()
    
    #plt.cla()

    #plt.close( fig )
    #plt.close( 'all' )
    
    del day2, day_diff, final, final1
    
    gc.collect()
    
    #print( 'clean_up: ' , time.time() - start__ )
    
    #print( 'total_iter: ' , time.time() - start__t )


    
    #h = hpy()
    #print(h.heap())

    
str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

# make a text_file of all the tiffs in the folder
text_file = open(str_name + '.txt', 'w')

files = sorted( os.listdir() )

for file in files:

    
    if 'tiff' in file:

        n = text_file.write('file ' + "'" + file + "'" + '\n')

text_file.close()

## command line script for turning the series of pictures produced in the script below to movies for upload

txt_name = str_name + '.txt'
vid_name = str_name + '.mp4'

command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
check_call(shlex.split(command))


to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )

str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

# make a text_file of all the tiffs in the folder
text_file = open(str_name + '.txt', 'w')

files = sorted( os.listdir() )

for file in files:

    
    if 'tiff' in file:

        n = text_file.write('file ' + "'" + file + "'" + '\n')

text_file.close()

## command line script for turning the series of pictures produced in the script below to movies for upload

txt_name = str_name + '.txt'
vid_name = str_name + '.mp4'

command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
check_call(shlex.split(command))


#to_delete = glob.glob('*.tiff')

#for item in to_delete:
#    os.remove( item )

to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )

os.chdir( 'F:/thermal_vids/2019-08-03/' )

str_name = '20190804_013007571000'

## command line script for turning the series of pictures produced in the script below to movies for upload

# make a text_file of all the tiffs in the folder
text_file = open(str_name + '.txt', 'w')

files = sorted( os.listdir() )

for file in files:

    
    if 'tiff' in file:

        n = text_file.write('file ' + "'" + file + "'" + '\n')

text_file.close()


txt_name = str_name + '.txt'
vid_name = str_name + '.mp4'

command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
check_call(shlex.split(command))


to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )

to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )

## writing all frames from a given thermal recording, separated into minute bins. Timestamps unrounded

zoom = 5

start_frame = 0

vid_length = 1*60*60

drive = 'D:/'

start_timestamp = _date_list_[ 1 ]

write_to = 'E:/thermal_vids/' + start_timestamp

#write_to = 'L:/analysis/annotation_vids/' + start_timestamp

start_dt = dt.datetime.strptime(start_timestamp, '%Y-%m-%d')

curr_date = f'{start_dt.year:04}' + f'{start_dt.month:02}' + f'{start_dt.day:02}'

os.chdir( drive + curr_date )

file_name = glob.glob('*.seq')

im = fnv.file.ImagerFile(file_name[0])

im.get_frame(0)

print( 'got_frame')

day1 = int(im.frame_info[0]['value'][:3])

start = f'{start_dt.year:04}-' + f'{start_dt.month:02}-' + f'{start_dt.day:02} ' + im.frame_info[0]['value'][4:]

start_dt = dt.datetime.strptime(start, '%Y-%m-%d %H:%M:%S.%f')

og_start_dt = start_dt


if not os.path.isdir( write_to + '/' ):
    
    os.makedirs ( write_to + '/')

os.chdir( write_to + '/')


final = np.array(im.final, copy=False).reshape((im.height, im.width))
    
width = final.shape[1] / (100 / zoom)
    
height = final.shape[0] / (100 / zoom)

fig = plt.figure( figsize = ( width, height ) )  

if start_frame > 0:
    
    counter = 0
    
else:
    
    counter = 1
    
for i in range(start_frame, im.num_frames):
    
    if im.get_frame( i ) == False:
        blaasdfadff
        
    #start__ = time.time()

    im.get_frame(i)

    #print( 'get_frame: ' , time.time() - start__ )

    #start__ = time.time()
    
    day2 = int(im.frame_info[0]['value'][:3])

    day_diff = day2 - day1

    timestamp = f'{og_start_dt.year:04}-' + f'{og_start_dt.month:02}-' + f'{(og_start_dt.day+day_diff):02} ' + im.frame_info[0]['value'][4:]

    timestamp = dt.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
    
    #print( 'timestamp_info: ' , time.time() - start__ )

    #start__t = time.time()
    
    if timestamp > ( start_dt + dt.timedelta( seconds = vid_length ) ):
        
        if not counter == 0:
           
            str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

            # make a text_file of all the tiffs in the folder
            text_file = open(str_name + '.txt', 'w')

            files = sorted( os.listdir() )

            for file in files:

                if 'tiff' in file:

                    n = text_file.write('file ' + "'" + file + "'" + '\n')

            text_file.close()

            ## command line script for turning the series of pictures produced in the script below to movies for upload

            txt_name = str_name + '.txt'
            vid_name = str_name + '.mp4'

            command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
            check_call(shlex.split(command))

        
        to_delete = glob.glob('*.tiff')
        
        for item in to_delete:
            os.remove( item )
        
        start_dt = timestamp
        
        counter = 1
        #os.makedirs ( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
        
        #os.chdir( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
     
    
    #print( 'if_statement: ' , time.time() - start__t )

    #start__ = time.time()
        
    final = np.array(im.final, copy=False).reshape( ( im.height, im.width ) )

    final1 = final - final.mean()
    
    #print( 'matrix: ' , time.time() - start__ )

    #start__ = time.time()

    sns.heatmap(final1, vmin = -np.std(final1), vmax = final1.max(), xticklabels = False, yticklabels = False, cbar = False)

    if final1.shape[1] > 1000:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 55*zoom / 5 )

    else:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 25*zoom / 5 )

    
    #print( 'plotting: ' , time.time() - start__ )

    #start__ = time.time()

    plt.savefig( f'{i:06}_' + timestamp.strftime("%Y%m%d_%H%M%S%f") + '.tiff' , bbox_inches = 'tight')
    
    #print( 'save_plot: ' , time.time() - start__ )

    #start__ = time.time()
    
    fig.clear()
    
    #plt.clf()
    
    #plt.cla()

    #plt.close( fig )
    #plt.close( 'all' )
    
    del day2, day_diff, final, final1
    
    gc.collect()
    
    #print( 'clean_up: ' , time.time() - start__ )
    
    #print( 'total_iter: ' , time.time() - start__t )


    
    #h = hpy()
    #print(h.heap())


str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

# make a text_file of all the tiffs in the folder
text_file = open(str_name + '.txt', 'w')

files = sorted( os.listdir() )

for file in files:

    
    if 'tiff' in file:

        n = text_file.write('file ' + "'" + file + "'" + '\n')

text_file.close()

## command line script for turning the series of pictures produced in the script below to movies for upload

txt_name = str_name + '.txt'
vid_name = str_name + '.mp4'

command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
check_call(shlex.split(command))


#to_delete = glob.glob('*.tiff')

#for item in to_delete:
#    os.remove( item )

to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )

## writing all frames from a given thermal recording, separated into minute bins. Timestamps unrounded

zoom = 5

start_frame = 0

vid_length = 1*60*60

drive = 'D:/'

start_timestamp = _date_list_[ 2 ]

write_to = 'E:/thermal_vids/' + start_timestamp

#write_to = 'L:/analysis/annotation_vids/' + start_timestamp

start_dt = dt.datetime.strptime(start_timestamp, '%Y-%m-%d')

curr_date = f'{start_dt.year:04}' + f'{start_dt.month:02}' + f'{start_dt.day:02}'

os.chdir( drive + curr_date )

file_name = glob.glob('*.seq')

im = fnv.file.ImagerFile(file_name[0])

im.get_frame(0)

print( 'got_frame')

day1 = int(im.frame_info[0]['value'][:3])

start = f'{start_dt.year:04}-' + f'{start_dt.month:02}-' + f'{start_dt.day:02} ' + im.frame_info[0]['value'][4:]

start_dt = dt.datetime.strptime(start, '%Y-%m-%d %H:%M:%S.%f')

og_start_dt = start_dt


if not os.path.isdir( write_to + '/' ):
    
    os.makedirs ( write_to + '/')

os.chdir( write_to + '/')


final = np.array(im.final, copy=False).reshape((im.height, im.width))
    
width = final.shape[1] / (100 / zoom)
    
height = final.shape[0] / (100 / zoom)

fig = plt.figure( figsize = ( width, height ) )  

if start_frame > 0:
    
    counter = 0
    
else:
    
    counter = 1
    
for i in range(start_frame, im.num_frames):
    
    if im.get_frame( i ) == False:
        blaasdfadff
        
    #start__ = time.time()

    im.get_frame(i)

    #print( 'get_frame: ' , time.time() - start__ )

    #start__ = time.time()
    
    day2 = int(im.frame_info[0]['value'][:3])

    day_diff = day2 - day1

    timestamp = f'{og_start_dt.year:04}-' + f'{og_start_dt.month:02}-' + f'{(og_start_dt.day+day_diff):02} ' + im.frame_info[0]['value'][4:]

    timestamp = dt.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
    
    #print( 'timestamp_info: ' , time.time() - start__ )

    #start__t = time.time()
    
    if timestamp > ( start_dt + dt.timedelta( seconds = vid_length ) ):
        
        if not counter == 0:
           
            str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

            # make a text_file of all the tiffs in the folder
            text_file = open(str_name + '.txt', 'w')

            files = sorted( os.listdir() )

            for file in files:

                if 'tiff' in file:

                    n = text_file.write('file ' + "'" + file + "'" + '\n')

            text_file.close()

            ## command line script for turning the series of pictures produced in the script below to movies for upload

            txt_name = str_name + '.txt'
            vid_name = str_name + '.mp4'

            command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
            check_call(shlex.split(command))

        
        to_delete = glob.glob('*.tiff')
        
        for item in to_delete:
            os.remove( item )
        
        start_dt = timestamp
        
        counter = 1
        #os.makedirs ( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
        
        #os.chdir( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
     
    
    #print( 'if_statement: ' , time.time() - start__t )

    #start__ = time.time()
        
    final = np.array(im.final, copy=False).reshape( ( im.height, im.width ) )

    final1 = final - final.mean()
    
    #print( 'matrix: ' , time.time() - start__ )

    #start__ = time.time()

    sns.heatmap(final1, vmin = -np.std(final1), vmax = final1.max(), xticklabels = False, yticklabels = False, cbar = False)

    if final1.shape[1] > 1000:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 55*zoom / 5 )

    else:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 25*zoom / 5 )

    
    #print( 'plotting: ' , time.time() - start__ )

    #start__ = time.time()

    plt.savefig( f'{i:06}_' + timestamp.strftime("%Y%m%d_%H%M%S%f") + '.tiff' , bbox_inches = 'tight')
    
    #print( 'save_plot: ' , time.time() - start__ )

    #start__ = time.time()
    
    fig.clear()
    
    #plt.clf()
    
    #plt.cla()

    #plt.close( fig )
    #plt.close( 'all' )
    
    del day2, day_diff, final, final1
    
    gc.collect()
    
    #print( 'clean_up: ' , time.time() - start__ )
    
    #print( 'total_iter: ' , time.time() - start__t )


    
    #h = hpy()
    #print(h.heap())


str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

# make a text_file of all the tiffs in the folder
text_file = open(str_name + '.txt', 'w')

files = sorted( os.listdir() )

for file in files:

    
    if 'tiff' in file:

        n = text_file.write('file ' + "'" + file + "'" + '\n')

text_file.close()

## command line script for turning the series of pictures produced in the script below to movies for upload

txt_name = str_name + '.txt'
vid_name = str_name + '.mp4'

command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
check_call(shlex.split(command))


#to_delete = glob.glob('*.tiff')

#for item in to_delete:
#    os.remove( item )

to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )

## writing all frames from a given thermal recording, separated into minute bins. Timestamps unrounded

zoom = 5

start_frame = 0

vid_length = 1*60*60

drive = 'D:/'

start_timestamp = _date_list_[ 3 ]

write_to = 'E:/thermal_vids/' + start_timestamp

#write_to = 'L:/analysis/annotation_vids/' + start_timestamp

start_dt = dt.datetime.strptime(start_timestamp, '%Y-%m-%d')

curr_date = f'{start_dt.year:04}' + f'{start_dt.month:02}' + f'{start_dt.day:02}'

os.chdir( drive + curr_date )

file_name = glob.glob('*.seq')

im = fnv.file.ImagerFile(file_name[0])

im.get_frame(0)

print( 'got_frame')

day1 = int(im.frame_info[0]['value'][:3])

start = f'{start_dt.year:04}-' + f'{start_dt.month:02}-' + f'{start_dt.day:02} ' + im.frame_info[0]['value'][4:]

start_dt = dt.datetime.strptime(start, '%Y-%m-%d %H:%M:%S.%f')

og_start_dt = start_dt

if not os.path.isdir( write_to + '/' ):
    
    os.makedirs ( write_to + '/')

os.chdir( write_to + '/')


final = np.array(im.final, copy=False).reshape((im.height, im.width))
    
width = final.shape[1] / (100 / zoom)
    
height = final.shape[0] / (100 / zoom)

fig = plt.figure( figsize = ( width, height ) )  

if start_frame > 0:
    
    counter = 0
    
else:
    
    counter = 1
    
for i in range(start_frame, im.num_frames):
    
    if im.get_frame( i ) == False:
        blaasdfadff
        
    #start__ = time.time()

    im.get_frame(i)

    #print( 'get_frame: ' , time.time() - start__ )

    #start__ = time.time()
    
    day2 = int(im.frame_info[0]['value'][:3])

    day_diff = day2 - day1

    timestamp = f'{og_start_dt.year:04}-' + f'{og_start_dt.month:02}-' + f'{(og_start_dt.day+day_diff):02} ' + im.frame_info[0]['value'][4:]

    timestamp = dt.datetime.strptime(timestamp, '%Y-%m-%d %H:%M:%S.%f')
    
    #print( 'timestamp_info: ' , time.time() - start__ )

    #start__t = time.time()
    
    if timestamp > ( start_dt + dt.timedelta( seconds = vid_length ) ):
        
        if not counter == 0:
           
            str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

            # make a text_file of all the tiffs in the folder
            text_file = open(str_name + '.txt', 'w')

            files = sorted( os.listdir() )

            for file in files:

                if 'tiff' in file:

                    n = text_file.write('file ' + "'" + file + "'" + '\n')

            text_file.close()

            ## command line script for turning the series of pictures produced in the script below to movies for upload

            txt_name = str_name + '.txt'
            vid_name = str_name + '.mp4'

            command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
            check_call(shlex.split(command))

        
        to_delete = glob.glob('*.tiff')
        
        for item in to_delete:
            os.remove( item )
        
        start_dt = timestamp
        
        counter = 1
        #os.makedirs ( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
        
        #os.chdir( write_to + '/' + start_dt.strftime("%Y%m%d_%H%M%S%f") )
     
    
    #print( 'if_statement: ' , time.time() - start__t )

    #start__ = time.time()
        
    final = np.array(im.final, copy=False).reshape( ( im.height, im.width ) )

    final1 = final - final.mean()
    
    #print( 'matrix: ' , time.time() - start__ )

    #start__ = time.time()

    sns.heatmap(final1, vmin = -np.std(final1), vmax = final1.max(), xticklabels = False, yticklabels = False, cbar = False)

    if final1.shape[1] > 1000:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 55*zoom / 5 )

    else:

        plt.xlabel(  timestamp.strftime("%Y-%m-%d %H:%M:%S.%f") , fontsize = 25*zoom / 5 )

    
    #print( 'plotting: ' , time.time() - start__ )

    #start__ = time.time()

    plt.savefig( f'{i:06}_' + timestamp.strftime("%Y%m%d_%H%M%S%f") + '.tiff' , bbox_inches = 'tight')
    
    #print( 'save_plot: ' , time.time() - start__ )

    #start__ = time.time()
    
    fig.clear()
    
    #plt.clf()
    
    #plt.cla()

    #plt.close( fig )
    #plt.close( 'all' )
    
    del day2, day_diff, final, final1
    
    gc.collect()
    
    #print( 'clean_up: ' , time.time() - start__ )
    
    #print( 'total_iter: ' , time.time() - start__t )


    
    #h = hpy()
    #print(h.heap())


str_name = start_dt.strftime("%Y%m%d_%H%M%S%f")

# make a text_file of all the tiffs in the folder
text_file = open(str_name + '.txt', 'w')

files = sorted( os.listdir() )

for file in files:

    
    if 'tiff' in file:

        n = text_file.write('file ' + "'" + file + "'" + '\n')

text_file.close()

## command line script for turning the series of pictures produced in the script below to movies for upload

txt_name = str_name + '.txt'
vid_name = str_name + '.mp4'

command = 'ffmpeg -f concat -r 60 -i ' + pipes.quote( txt_name ) + ' -c:v libx264 -crf 15 -pix_fmt yuv420p -vf pad=ceil(iw/2)*2:ceil(ih/2)*2 ' + pipes.quote(vid_name)
check_call(shlex.split(command))


#to_delete = glob.glob('*.tiff')

#for item in to_delete:
#    os.remove( item )

to_delete = glob.glob('*.tiff')

for item in to_delete:
    os.remove( item )
