==> CODE/07_track_correction_GUI.ipynb <==
#import
import cv2
from TrackCorrectionGUI import TrackCorrectionGUI

#SETUP

## I need to start with track 4!

# The yellow track is the focal track that is being corrected
# The blue track is the other track that might be merged with 
# the focal track

# KEY FUNCTIONS:
# next frame = "c"
# previous frame = "y"
# advance 3 frames = "d"
# go back 3 frames = "a"
# advance 30 frames = "e"
# go back 30 frames = "q"
# advance 90 frames = "3"
# go back 90 frames = "1"
# advance 500 frames = "6"
# go back 500 frames = "5"

# go to frame at beginning of current possible merge track = "b"
# go to frame at end of yellow (focal) track = "g"
# next possible merge track = "w"
# previous possible merge track = "x"
# next focal track = "." (same button as '>')
# previous focal track = "," (same button as '<')
# add point to focal track = click anywhere on picture
# merge focal track with current selected merge track = space bar
# permanently remove focal track = "ä" 

# split focal track at current frame = "k"
# split possible merge track at current frame = "l"

# undo last action = "u"
# hide tracks (so easier to see frame) = "h"

# to add a new track = hold down the shift key and then click anywhere on the picture

# add a point to a track WITHOUT linear interpolation = hold down the control key and click anywhere on the picture

#THINGS FOR USER TO ADD:
#file containing info about the tracks

# The tracks file should be a list of dictionaries]
# Each dictionary contains information about one track
# The dictionary must have the key "track", "first_frame", "last_frame"
# the key "track" returns an numpy array of shape (N, 2) where N is the length of the track in frames
# each point in the track is in numpy array indexing so row, column from top left corner of image

# The key "first_frame" returns the first frame number of the video the track appears
# The key "last_frame" returns the last frame number of the video that the track appears


tracks_file = 'C:/Users/meerkat/Documents/thermal_baboon/RESULTS/tracking_output/20190802/tracks_20190802_180507743000-final-final.npy'
#tracks_file_demo = 'C:/Users/meerkat/Documents/thermal_baboon/CODE/manual_track_correction/demo-correction-tracks-xy.npy'
    
# Path to frames
frames_folder = 'C:/Users/meerkat/Documents/thermal_baboon/DATA/frames_for_finalizing_Vilson_1/'
    
# how much you want to rescale the original frame resolution to display (and fit)on the screen    
factor = 0.62

import numpy as np

tracks = np.load(tracks_file, allow_pickle=True)

#tracks_demo = np.load(tracks_file_demo, allow_pickle=True)

## tracks file correction

for t in range( len( tracks ) ):
    
    tracks[ t ]['pos_index'] = tracks[ t ]['pos_index'].reshape( tracks[ t ]['pos_index'].shape[0], 1 )

    
np.save( tracks_file, tracks  )



import glob
import os

frame_files = sorted(glob.glob(os.path.join(frames_folder, '*.jpg')))

#RUN

#get good track, draw image on window, set mouse to work
window = TrackCorrectionGUI(tracks_file, frame_files, factor, point_scale=1, xy_tracks=True)


window.find_next_track()
window.draw_window()
cv2.namedWindow('pic0')
cv2.setMouseCallback('pic0', window.clicked)

#loop to keep image updating
while True:
    # exit when user presses esc key
    key = cv2.waitKey(2) & 0xff 
    window.detect_keys(key)
    if key == 27:
        window.save('active')
        break
cv2.destroyAllWindows()

##### len( frame_files )
