==> CODE/10_positions_npy_to_csv.ipynb <==
import pandas as pd, numpy as np, cv2, os, glob, matplotlib.pyplot as plt

input_dir = 'C:/Users/meerkat/Documents/social_sleep/DATA/tracking_output/more_tracks_parameters/'

os.chdir( input_dir )

dirs_to_process = os.listdir()

os.chdir( os.path.join( input_dir, dirs_to_process[ 0 ] ))

positions_files = glob.glob( 'positions*.npy' )

for direc in dirs_to_process:
    
    print( direc )
    
    os.chdir( os.path.join( input_dir, direc ) )
    
    positions_files = glob.glob( 'positions*.npy' )
    
    for pos_file in positions_files:
        
        positions_filepath = os.path.join( input_dir, direc, pos_file )

        positions = np.load( positions_filepath, allow_pickle = True )
        
        # make an dataframe with 0 rows
        all_positions = pd.DataFrame({'frame':[],  'x':[],  'y':[] }) # start a dataframe that will be filled with all the location of all the tracks for the video
        
        # fill the dataframe by appending rows
        for frame_num, pos_arr in enumerate( positions ):
    
            if pos_arr.shape[ 0 ] > 0:

                to_append = pd.DataFrame( {'frame': pd.Series( [ frame_num ] * pos_arr.shape[ 0 ] ), 'x': pd.Series( pos_arr[:,0] ) ,  'y': pd.Series( pos_arr[:,1] ) } )    

                all_positions = all_positions.append( to_append, ignore_index = True ) 

        vid_name  = positions_filepath.split( "/")[ -1 ].split( "_", 1 )[ 1 ].split( "." )[ 0 ]

        all_positions[ 'vid_name' ] =  vid_name

        all_positions = all_positions.astype( { 'frame': int, 'x': float, 'y': float, 'vid_name': str } )
        
        all_positions.to_csv( os.path.join( input_dir, direc, 'positions_' + vid_name + '.csv' ), index=False) # write the csv with all the tracks to a file
