==> CODE/13_parsing_acc_for_acc_gps_thermal_for_ID.ipynb <==
import subprocess
import os
import glob
import sys
import time
import pandas as pd
import numpy as np
import datetime as dt
from sortedcontainers import SortedSet
import gc

from decord import VideoReader, cpu, gpu
import cv2
import subprocess as sp
from tqdm import tqdm

import matplotlib.pyplot as plt

### load in the full acc data

os.chdir("C:/Users/meerkat/Documents/social_sleep/DATA/gps_acc_data/")

acc_dat = pd.read_csv('acc_dat.csv', header=0)

## only keep the necessary columns
acc_dat = acc_dat[ ['tag', 'corr_local_timestamp', 'acc_samp_freq', 'acc'] ]

# turn the timestamp into a datetime element
acc_dat[ 'corr_local_timestamp' ] = pd.to_datetime( acc_dat[ 'corr_local_timestamp' ] )

# save the path to the folder containing the night folders which contain .txt files with their respective processed videos
root = 'Z:/baboon/archive/processed/video/thermal/2019_summer/cliff_data/mp4_files/viewpoint_1/T1020'


# change the working directory to this folder
os.chdir( root )

# make a list of the night folders within this directory
nights = os.listdir()

# save the tag names
tag_names = SortedSet( acc_dat[ 'tag'] )

# this is the time in seconds of acc that you want to keep before a video starts and after a video ends
buffer = dt.timedelta( seconds = 120 ) 

# for each night
for night in nights:
    
    # save the path to this night's folder
    folder_path = os.path.join( root, night )
    
    # change the working directory to this night's folder
    os.chdir( folder_path )
    
    # make a list of the .txt files in the folder
    txt_files = glob.glob( "*.txt" )
    
    for txt_file in txt_files:
        
        hour_name = txt_file.split( '.' )[ 0 ]
        
        txt_filename = os.path.join( folder_path, txt_file )
        
        with open(txt_filename) as f:
            lines = f.readlines()
    
        frame_timestamps = list( map( lambda x: x.split( "'")[ 1 ].split( '.' ) [ 0 ].split( "_", 1 )[ 1 ] , lines ) )
        frame_timestamps = list( map( lambda x: dt.datetime.strptime(x, '%Y%m%d_%H%M%S%f'), frame_timestamps ) ) 

        frame_local_timestamps = list( map( lambda x: x + dt.timedelta( hours = 3 ), frame_timestamps ) )

        start_time = frame_local_timestamps[ 0 ]

        end_time = frame_local_timestamps[ ( len( frame_local_timestamps ) - 1 ) ]


        # these should be the same
        print( hour_name )
        print( start_time )

        # trim the acc data to the time of this video (plus the buffer before the start of the video and the buffer after the end of the video)
        sub_ved = acc_dat[ ( acc_dat[ 'corr_local_timestamp' ] > ( start_time - buffer ) ) & ( acc_dat[ 'corr_local_timestamp' ] < ( end_time + buffer ) ) ]
        
        # make a directory
        os.makedirs( 'C:/Users/meerkat/Documents/thermal_baboon/DATA/gps_acc_data/parsed_acc_for_animation/' + night + '/' + hour_name + '/', exist_ok = True )

        os.chdir( 'C:/Users/meerkat/Documents/thermal_baboon/DATA/gps_acc_data/parsed_acc_for_animation/' + night + '/' + hour_name + '/' )

        for tag in tag_names:

            print( tag )

            tag_dat = sub_ved[ sub_ved[ 'tag' ] == tag ]

            parsed_dat = pd.DataFrame( [] )

            for i in range( tag_dat.shape[ 0 ] ):

                if tag_dat.iloc[ i, 2 ] == 12:

                    to_add = pd.DataFrame( {'tag': tag, 'corr_local_timestamp': pd.date_range( tag_dat.iloc[ i, 1 ], tag_dat.iloc[ i, 1 ] + dt.timedelta( seconds = 1 ), 13 )[ :12 ], 'x': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 0 :: 3 ], 'y': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 1 :: 3 ], 'z': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 2 :: 3 ] } ) 

                    parsed_dat = pd.concat( [ parsed_dat, to_add ] )

                elif tag_dat.iloc[ i, 2 ] == 20:

                    to_add = pd.DataFrame( {'tag': tag, 'corr_local_timestamp': pd.date_range( tag_dat.iloc[ i, 1 ], tag_dat.iloc[ i, 1 ] + dt.timedelta( seconds = 2 ), 41 )[ :40 ], 'x': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 0 :: 3 ], 'y': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 1 :: 3 ], 'z': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 2 :: 3 ] } ) 

                    parsed_dat = pd.concat( [ parsed_dat, to_add ] )

                elif tag_dat.iloc[ i, 2 ] == 56.23:

                    to_add = pd.DataFrame( {'tag': tag, 'corr_local_timestamp': pd.date_range( tag_dat.iloc[ i, 1 ], tag_dat.iloc[ i, 1 ] + dt.timedelta( seconds = 1 ), 56.23 )[ :40 ], 'x': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 0 :: 3 ], 'y': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 1 :: 3 ], 'z': tag_dat[ 'acc' ].iloc[ i ].split( ' ' )[ 2 :: 3 ] } ) 

                    parsed_dat = pd.concat( [ parsed_dat, to_add ] )

            parsed_dat.to_csv( str( tag ) + '_' + hour_name + '.csv' )
